Move handshake_close() function into connection module
