+++ quinn-proto/src/connection.rs
@@ -2487,6 +2487,45 @@ pub fn parse_initial(log: &Logger, payload: Bytes) -> Result<Option<frame::Strea
     Ok(result)
 }
 
+pub fn handshake_close<R>(
+    crypto: &Crypto,
+    remote_id: &ConnectionId,
+    local_id: &ConnectionId,
+    packet_number: u32,
+    reason: R,
+    tls_alert: Option<&[u8]>,
+) -> Box<[u8]>
+where
+    R: Into<state::CloseReason>,
+{
+    let mut buf = Vec::<u8>::new();
+    Header::Long {
+        ty: LongType::Handshake,
+        dst_cid: *remote_id,
+        src_cid: *local_id,
+        number: packet_number,
+    }.encode(&mut buf);
+    let header_len = buf.len();
+    let max_len = MIN_MTU - header_len as u16 - AEAD_TAG_SIZE as u16;
+    match reason.into() {
+        state::CloseReason::Application(ref x) => x.encode(&mut buf, max_len),
+        state::CloseReason::Connection(ref x) => x.encode(&mut buf, max_len),
+    }
+    if let Some(data) = tls_alert {
+        if !data.is_empty() {
+            frame::Stream {
+                id: StreamId(0),
+                fin: false,
+                offset: 0,
+                data,
+            }.encode(false, &mut buf);
+        }
+    }
+    set_payload_length(&mut buf, header_len);
+    crypto.encrypt(packet_number as u64, &mut buf, header_len);
+    buf.into()
+}
+
 /// Reasons why a connection might be lost.
 #[derive(Debug, Clone, Fail)]
 pub enum ConnectionError {
